veriCekme/requirements.txt
request==2.25.1
bs4==4.9.3